+++ src/RadixDlt.NetworkGateway.PostgresIntegration/Migrations/20241011144531_InitialCreate.Designer.cs
@@ -81,7 +81,7 @@ using RadixDlt.NetworkGateway.PostgresIntegration.Models;
 namespace RadixDlt.NetworkGateway.PostgresIntegration.Migrations
 {
     [DbContext(typeof(MigrationsDbContext))]
-    [Migration("20241011143716_InitialCreate")]
+    [Migration("20241011144531_InitialCreate")]
     partial class InitialCreate
     {
         /// <inheritdoc />
@@ -273,6 +273,34 @@ namespace RadixDlt.NetworkGateway.PostgresIntegration.Migrations
                     b.ToTable("account_locker_entry_resource_vault_definition");
                 });
 
+            modelBuilder.Entity("RadixDlt.NetworkGateway.PostgresIntegration.Models.AccountLockerEntryResourceVaultTotalsHistory", b =>
+                {
+                    b.Property<long>("Id")
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("bigint")
+                        .HasColumnName("id");
+
+                    NpgsqlPropertyBuilderExtensions.UseIdentityByDefaultColumn(b.Property<long>("Id"));
+
+                    b.Property<long>("AccountLockerDefinitionId")
+                        .HasColumnType("bigint")
+                        .HasColumnName("account_locker_definition_id");
+
+                    b.Property<long>("FromStateVersion")
+                        .HasColumnType("bigint")
+                        .HasColumnName("from_state_version");
+
+                    b.Property<long>("TotalResources")
+                        .HasColumnType("bigint")
+                        .HasColumnName("total_resources");
+
+                    b.HasKey("Id");
+
+                    b.HasIndex("AccountLockerDefinitionId", "FromStateVersion");
+
+                    b.ToTable("account_locker_entry_resource_vault_totals_history");
+                });
+
             modelBuilder.Entity("RadixDlt.NetworkGateway.PostgresIntegration.Models.AccountLockerEntryTouchHistory", b =>
                 {
                     b.Property<long>("Id")
@@ -295,6 +323,34 @@ namespace RadixDlt.NetworkGateway.PostgresIntegration.Migrations
                     b.HasIndex("AccountLockerDefinitionId", "FromStateVersion");
 
                     b.ToTable("account_locker_entry_touch_history");
+                });
+
+            modelBuilder.Entity("RadixDlt.NetworkGateway.PostgresIntegration.Models.AccountLockerTotalsHistory", b =>
+                {
+                    b.Property<long>("Id")
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("bigint")
+                        .HasColumnName("id");
+
+                    NpgsqlPropertyBuilderExtensions.UseIdentityByDefaultColumn(b.Property<long>("Id"));
+
+                    b.Property<long>("AccountLockerEntityId")
+                        .HasColumnType("bigint")
+                        .HasColumnName("account_locker_entity_id");
+
+                    b.Property<long>("FromStateVersion")
+                        .HasColumnType("bigint")
+                        .HasColumnName("from_state_version");
+
+                    b.Property<long>("TotalAccounts")
+                        .HasColumnType("bigint")
+                        .HasColumnName("total_accounts");
+
+                    b.HasKey("Id");
+
+                    b.HasIndex("AccountLockerEntityId", "FromStateVersion");
+
+                    b.ToTable("account_locker_totals_history");
                 });
 
             modelBuilder.Entity("RadixDlt.NetworkGateway.PostgresIntegration.Models.AccountResourcePreferenceRuleAggregateHistory", b =>